django_poll_api/requirements.txt
appdirs==1.4.4
certifi==2020.6.20
chardet==3.0.4
conda==4.3.16
distlib==0.3.1
Django==2.2.10
django-debug-toolbar==3.2
djangorestframework==3.12.2
filelock==3.0.12
idna==2.10
pipenv==2020.8.13
pycosat==0.6.3
pytz==2020.5
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
sqlparse==0.4.1
urllib3==1.25.10
virtualenv==20.0.31
virtualenv-clone==0.5.4
